BRD-SalesCRM-Dashboard-Backend/requirements.txt
# Django REST Framework Sales CRM Dashboard Backend
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.8
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.10.1

# Development & Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Date & Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# File Handling & Media
Pillow==10.0.1
django-storages==1.13.2

# Email & Communication
django-anymail==10.2
celery==5.3.4
redis==5.0.1

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# API & HTTP Requests
requests==2.31.0
urllib3==2.0.7

# Validation & Serialization
marshmallow==3.20.1
jsonschema==4.19.2

# Logging & Monitoring
sentry-sdk==1.32.0
django-debug-toolbar==4.2.0

# Performance & Optimization
django-cacheops==7.0.2
django-redis==5.3.0

# Utilities
click==8.1.7
rich==13.6.0
tqdm==4.66.1

# Production Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Development Tools
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Optional: For Excel/CSV export
xlsxwriter==3.1.9
tablib==3.5.0

# Optional: For PDF generation
reportlab==4.0.7
WeasyPrint==60.0

# Optional: For advanced caching
memcached==1.62

# Optional: For background tasks
flower==2.0.1

# Optional: For API testing
httpie==3.2.2
postman==1.0.0
